% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{drre_avg}
\alias{drre_avg}
\title{drre_avg: Average Effects of Covariates Based on `drre' fit}
\usage{
drre_avg(t1 = t1, t2 = t2, result = estimated_effects)
}
\arguments{
\item{t1}{Lower limit of the time interval of interest}

\item{t2}{Upper limit of the time interval of interest}

\item{result}{Data-frame or matrix argument of the vector of unique event times in the sample (increasing order) and the estimated effects of covariates over the event times, i.e., the drre fit}
}
\description{
Function `drre_ave' provides an average of the estimated effects of covariates coefficients discussed in Soh and Huang (Biometrics, 2019; Section 3.4); see Peng and Huang (1981) for its usage in quantile regression
}
\examples{
drre_avg()
}
\keyword{dynamic}
\keyword{events;}
\keyword{frequency}
\keyword{function;}
\keyword{mean}
\keyword{recurrent}
\keyword{regression;}
