% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{drre_TS}
\alias{drre_TS}
\title{drre_TS: A Test Statistic for Evaluating H_0: the effect of a covariate is constant over (t1, t2]}
\usage{
drre_TS(t1 = t1, t2 = t2, result = estimated_effects, pwr = pwr,
  n = n)
}
\arguments{
\item{t1}{Lower limit of the time interval of interest}

\item{t2}{Upper limit of the time interval of interest}

\item{result}{Data-frame or matrix argument of the vector of unique event times in the sample (increasing order) and the estimated effects of covariates over the event times, i.e., the drre fit}

\item{pwr}{The degree of a polynomial weight function}

\item{n}{the number of subjects in the sample}
}
\description{
Function `drre_TS' provides a test statistic for evaluating the null hypothesis H_0: the effect of a covariate is constant over (t1, t2], which is discussed in Soh and Huang (Biometrics, 2019; Section 3.4).
}
\examples{
drre_avg()
}
\keyword{dynamic}
\keyword{events;}
\keyword{frequency}
\keyword{function;}
\keyword{mean}
\keyword{recurrent}
\keyword{regression;}
