% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{drre_avg}
\alias{drre_avg}
\title{drre_avg: Average Effects of Covariates Based on `drre' fit}
\usage{
drre_avg(t1 = t1, t2 = t2, result = estimated_effects)
}
\arguments{
\item{t1}{Lower limit of the time interval of interest}

\item{t2}{Upper limit of the time interval of interest}

\item{result}{Data-frame or matrix argument of the vector of unique event times in the sample (increasing order) and the estimated effects of covariates over the event times, i.e., the drre fit}
}
\description{
Function `drre_ave' provides an average of the estimated effects of covariates coefficients discussed in Soh and Huang (Biometrics; Section 3.4); see Peng and Huang (1981) for its usage in quantile regression
}
\examples{
# ------------------
# Calling the bladder tumor data (Byar, 1980) from `survival' R package
# ------------------
require(survival)
data("bladder")

# ------------------
# Data Step
# ------------------
N=length(unique(bladder1$id))
uniq.evnts=sort(unique(bladder1[bladder1$status==1,"stop"]))

temp.cen=do.call("rbind",by(bladder1,bladder1$id,tail,n=1))[,c("status","stop")]
censoring=ifelse(temp.cen[,"status"]==1,temp.cen[,"stop"]+.1,temp.cen[,"stop"])

temp.cov=bladder1[!duplicated(bladder1$id),c("treatment","number","size")]
covar=cbind(rep(1,length(unique(bladder1$id))),
            as.numeric(temp.cov$treatment=="pyridoxine"),
            as.numeric(temp.cov$treatment=="thiotepa"),
            temp.cov$number,
            temp.cov$size
)

idMAT=matrix(0,nrow=length(unique(bladder1[bladder1$status==1,"stop"]))
             , ncol =  max(table(bladder1[bladder1$status==1,"stop"]))
)
temp.cnt=1
for(i in sort(unique(bladder1[bladder1$status==1,"stop"]))){
  temp=sort(bladder1[bladder1$status==1 & bladder1[,"stop"]==i,"id"])
  for (j in 1:length(temp)){
    idMAT[temp.cnt,j]=temp[j]
  }
  temp.cnt=temp.cnt+1
}

# ------------------
# Model Fitting
# ------------------
BETA1=
     drre(N,
     length(unique(bladder1[bladder1$status==1,"stop"])),
     4,
     max(table(bladder1[bladder1$status==1,"stop"])),
     uniq.evnts,
     censoring,
     as.vector(table(bladder1[bladder1$status==1,"stop"])),
     as.vector(t(covar)),
     as.vector(t(idMAT))
)
trim.result=cbind(uniq.evnts,BETA1[[1]][-1,])

# ------------------
# Estimates of Average Effects of Covariates in time interval (t1,t2]
# ------------------
t1=5; t2=53
drre_avg(t1,t2,trim.result)[-1]

}
\keyword{dynamic}
\keyword{events;}
\keyword{frequency}
\keyword{function;}
\keyword{mean}
\keyword{recurrent}
\keyword{regression;}
