% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{drre_mB}
\alias{drre_mB}
\title{drre_mB: `drre' Based on Nonparametric Multiplier Bootstrap}
\usage{
drre_mB(n, M, P, MCol, u, dt, MNum, zvec, Mvec, xiN)
}
\arguments{
\item{n}{the number of subjects}

\item{M}{the number of unique event times in the sample}

\item{P}{the number of covariates of interest}

\item{MCol}{the maximum number of subjects observed in the sample at an event time}

\item{u}{a vector of ordered unique event times in the sample}

\item{dt}{a vector of censoring times}

\item{zvec}{vectorized form of [1 Z]^T matrix. (5 x N)}

\item{Mvec}{vectorized form of (MCOL x # unique times) ID matrix. ids for unique event times}

\item{xiN}{vector of a random sample of size n which following the standard exponential distribution (mean=variance=1)}

\item{NNum}{the number of observations at unique event times; if no tied events, all is 0}
}
\description{
Function `drre_mB' fits the regression coefficients in the model based on the proposed nonparametric multiplier bootstrap estimation procedure in Soh and Huang (Biometrics, 2019); see related discussion in Rubin (1981), Kosorok (2008) and Huang (2014)
}
\examples{
drre_mB()
}
\keyword{dynamic}
\keyword{events;}
\keyword{frequency}
\keyword{function;}
\keyword{mean}
\keyword{recurrent}
\keyword{regression;}
